use restore keys to restore cache from partial key matches
From the GitHub Actions documentation:
https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#matching-a-cache-key

You can provide a list of restore keys to use when there is a
cache miss on key. You can create multiple restore keys ordered
from the most specific to least specific. The cache action
searches for restore-keys in sequential order. When a key doesn't
match directly, the action searches for keys prefixed with the
restore key. If there are multiple partial matches for a restore
key, the action returns the most recently created cache.